consolidate types in root
