+++ public/src/utils/initUI.ts
@@ -1,6 +1,6 @@
 import { state } from "../initialState.js";
 import { GUI_GUTTER, GUI_OFFSET, paintProperties as p } from "../constants.js";
-import { BrushShape, GuiValues, LfoValues } from "../types/paint";
+import { BrushShape, GuiValues, LfoValues } from "../../../types";
 import { PaintWithFriends } from "../app.js";
 
 export type p5GuiInstance = {
