+++ types.ts
@@ -1,5 +1,3 @@
-import { DrawUpdate } from "./websocket";
-
 export type BrushShape = "line" | "circle" | "square" | "text";
 
 export interface Paintbrush {
@@ -111,3 +109,46 @@ export type SetupShapeReturnValues = {
 export type MirrorModeParams = SetupShapeReturnValues & {
   fn: p5["line"] | p5["circle"] | p5["square"] | p5["text"];
 };
+
+export interface DrawUpdate
+  extends Omit<Paintbrush, "fillColor" | "strokeColor"> {
+  prevX: number;
+  prevY: number;
+  text: string;
+  fillColor: string;
+  strokeColor: string;
+}
+
+export type LeanDrawUpdate = [
+  string,
+  number,
+  number,
+  string,
+  number,
+  string,
+  number,
+  boolean,
+  boolean,
+  BrushShape,
+  number,
+  number,
+  number,
+  number,
+  number,
+  number,
+  string
+];
+
+interface Message {
+  sender: string;
+  message: string;
+  id: string;
+}
+
+export type Messages = Message[];
+
+export interface Connections {
+  [id: string]: string;
+}
+
+export type Payload = LeanDrawUpdate | Messages | Connections;
